pyproject.toml
[tool.poetry]
name = "study-flask-admin"
version = "0.1.0"
description = ""
authors = ["Leo Graciano <leonardoferreiragraciano@gmail.com>"]
readme = "README.md"
packages = [{include = "study_flask_admin"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
flask-admin = "^1.6.0"
flask-login = "^0.6.2"
flask-sqlalchemy = "^3.0.2"
python-dotenv = "^0.21.0"
python-decouple = "^3.6"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
